+++ Moco/Moco/MocoCasADiSolver/MocoCasOCProblem.cpp
@@ -55,16 +55,20 @@ MocoCasOCProblem::MocoCasOCProblem(const MocoCasADiSolver& mocoCasADiSolver,
             stateType = CasOC::StateType::Speed;
         else
             stateType = CasOC::StateType::Auxiliary;
-        addState(stateName, stateType, convertBounds(info.getBounds()),
-                convertBounds(info.getInitialBounds()),
-                convertBounds(info.getFinalBounds()));
+        if (info.getUseFunctionBounds()) {
+            addStateWithTimeVaryingBounds(stateName, stateType);
+        } else {
+            addState(stateName, stateType, convertBounds(info.getPhaseBounds()),
+                    convertBounds(info.getInitialBounds()),
+                    convertBounds(info.getFinalBounds()));
+        }
     }
 
     auto controlNames =
             createControlNamesFromModel(model, m_modelControlIndices);
     for (const auto& controlName : controlNames) {
         const auto& info = problemRep.getControlInfo(controlName);
-        addControl(controlName, convertBounds(info.getBounds()),
+        addControl(controlName, convertBounds(info.getPhaseBounds()),
                 convertBounds(info.getInitialBounds()),
                 convertBounds(info.getFinalBounds()));
     }
@@ -169,7 +173,7 @@ MocoCasOCProblem::MocoCasOCProblem(const MocoCasADiSolver& mocoCasADiSolver,
                     }
 
                     addKinematicConstraint(multInfo.getName(),
-                            convertBounds(multInfo.getBounds()),
+                            convertBounds(multInfo.getPhaseBounds()),
                             convertBounds(multInfo.getInitialBounds()),
                             convertBounds(multInfo.getFinalBounds()), kinLevel);