+++ Moco/Moco/MocoTrack.cpp
@@ -47,6 +47,10 @@ void MocoTrack::constructProperties() {
     constructProperty_allow_unused_references(false);
     constructProperty_guess_file("");
     constructProperty_apply_tracked_states_to_guess(false);
+    constructProperty_bound_kinematic_states();
+    constructProperty_state_bound_range_rotational(
+            SimTK::convertDegreesToRadians(20));
+    constructProperty_state_bound_range_translational(0.20);
     constructProperty_minimize_control_effort(true);
     constructProperty_control_effort_weight(0.001);
 }
@@ -101,6 +105,34 @@ MocoStudy MocoTrack::initialize() {
         m_timeInfo.final -= 1e-3;
     }
     problem.setTimeBounds(m_timeInfo.initial, m_timeInfo.final);
+
+    // Bound kinematic states.
+    // -----------------------
+    bool boundKinematicStates = false;
+    if (getProperty_bound_kinematic_states().empty()) {
+        if (!get_states_reference().empty()) {
+            boundKinematicStates = true;
+        }
+    } else {
+        OPENSIM_THROW_IF_FRMOBJ(
+                get_bound_kinematic_states() && get_states_reference().empty(),
+                Exception,
+                "Cannot bound kinematic states with states reference if no "
+                "states reference is supplied.");
+    }
+
+    checkPropertyIsNonnegative(
+            *this, getProperty_state_bound_range_rotational());
+    checkPropertyIsNonnegative(
+            *this, getProperty_state_bound_range_translational());
+
+    if (boundKinematicStates) {
+        setKinematicStateFunctionBoundsFromTable(model,
+                tracked_states,
+                get_state_bound_range_rotational(),
+                get_state_bound_range_translational(), problem);
+    }
+
 
     // Configure solver.
     // -----------------